[#58] move stores under src/shared folder
