+++ src/shared/lib/use-api-transport/use-api-transport.ts
@@ -1,8 +1,7 @@
 import type { RayContentLock } from "~/src/entities/ray/types";
+import { useEventStore, useConnectionStore } from "../../stores";
 import type { EventId, EventType } from '../../types';
 import { useCentrifuge, useEventsRequests } from "../io";
-import { useConnectionStore } from "~/stores/connections";
-import { useEventStore } from "~/stores/events";
 
 const CHECK_CONNECTION_INTERVAL = 10000
 export const useApiTransport = () => {
